requirements.txt
dgl
dgllife
easydict
pytorch-geometric
pytorch
ase
rdkit
chgnet
dscribe
hyperopt
joblib
matplotlib
numpy
pickle5
ray
scikit-learn
scipy
tensorboardX
seaborn